add: 複数選択終了

+++ src/database/migrations/2022_06_24_001226_create_records_table.php
@@ -16,7 +16,7 @@ class CreateRecordsTable extends Migration
         Schema::create('records', function (Blueprint $table) {
             $table->increments('id');
             $table->string('study_date');
-            $table->integer('study_time');
+            $table->decimal('study_time');
             $table->integer('language_id');
             $table->integer('content_id');
             $table->timestamps();
